+++ src/ABC_Util.h
@@ -77,7 +77,7 @@ extern "C" {
                                                     type *p;  \
                                                     type *end;  \
                                             }
-#define ABC_BUF_SIZE(x)                     (((x).end)-((x).p))
+#define ABC_BUF_SIZE(x)                     ((unsigned int) (((x).end)-((x).p)))
 #define ABC_BUF_FREE(x)                     {  \
                                                 free((x).p);  \
                                                 (x).p = NULL;  \
@@ -104,7 +104,7 @@ extern "C" {
                                                 (buf).end = (buf).p +      __abc_buf_expand_count__; \
                                             }
 #define ABC_BUF_APPEND_PTR(buf, ptr, count) { \
-                                                int __abc_buf_append_resize__ = (((buf).end)-((buf).p)) + (sizeof(*((buf).p)) * count); \
+                                                int __abc_buf_append_resize__ = (int) (((buf).end)-((buf).p)) + (sizeof(*((buf).p)) * count); \
                                                 (buf).p = realloc((buf).p, __abc_buf_append_resize__); \
                                                 (buf).end = (buf).p +      __abc_buf_append_resize__; \
                                                 memcpy((buf).end - count, ptr, count); \
@@ -117,13 +117,13 @@ extern "C" {
                                                 memcpy((dst).end - __abc_buf_append_size__, (src).p,  __abc_buf_append_size__); \
                                             }
 #define ABC_BUF_DUP(dst, src)               { \
-                                                int __abc_buf_dup_size__ = (((src).end)-((src).p)); \
+                                                int __abc_buf_dup_size__ = (int) (((src).end)-((src).p)); \
                                                 (dst).p = malloc(__abc_buf_dup_size__); \
                                                 (dst).end = (dst).p + __abc_buf_dup_size__; \
                                                 memcpy((dst).p, (src).p, __abc_buf_dup_size__); \
                                             }
 #define ABC_BUF_DUP_PTR(buf, ptr, size)     { \
-                                                int __abc_buf_dup_size__ = size; \
+                                                int __abc_buf_dup_size__ = (int) size; \
                                                 (buf).p = malloc(__abc_buf_dup_size__); \
                                                 (buf).end = (buf).p + __abc_buf_dup_size__; \
                                                 memcpy((buf).p, ptr, __abc_buf_dup_size__); \
